server/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.11.1
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
confuse==2.1.0
et_xmlfile==2.0.0
fastapi==0.122.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
nodeenv==1.9.1
numpy==2.3.5
openpyxl==3.1.5
pandas==2.3.3
prisma==0.15.0
pydantic==2.12.5
pydantic_core==2.41.5
python-amazon-sp-api==1.9.57
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
tomlkit==0.13.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.6.1
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
